Time: 0 ms (100%), Space: 89.5 MB (44.2%) - LeetHub

+++ 2270-number-of-ways-to-split-array/2270-number-of-ways-to-split-array.cpp
@@ -2,13 +2,13 @@ class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
         
-        int sum = 0;
+        long long sum = 0;
         for (int i = 0; i < nums.size(); i++) {
             sum += nums[i];
         }
 
         int result = 0; 
-        int runningSum = 0; 
+        long long runningSum = 0; 
         for (int i = 0; i < nums.size() - 1; i++) {
             if ((runningSum += nums[i]) >= sum - runningSum) result++;
         }
